Tank canyon stage + airplane stage

+++ Ragdoll Race/Assets/Stages/_Common/Knockout Trigger/KnockoutTrigger.cs
@@ -4,6 +4,9 @@ using UnityEngine;
 
 public class KnockoutTrigger : MonoBehaviour
 {
+    public bool knockoutOnEnter = true;
+    public bool knockoutOnExit = false;
+
     public Collider collider;
 
     private void Awake()
@@ -17,12 +20,42 @@ public class KnockoutTrigger : MonoBehaviour
     
     private void OnTriggerEnter(Collider other)
     {
-        Hittable hittable = other.gameObject.GetComponent<Hittable>();
+        if(knockoutOnEnter){
+            Hittable hittable = other.gameObject.GetComponent<Hittable>();
 
-        if(hittable){
-            Player player = hittable.player;
+            if(hittable){
+                Player player = hittable.player;
 
-            player.Kill();
+                player.Kill();
+            }
+        }   
+    }
+
+    private void OnCollisionEnter(Collision other)
+    {
+        if(knockoutOnEnter){
+            Hittable hittable = other.gameObject.GetComponent<Hittable>();
+
+            if(hittable){
+                Player player = hittable.player;
+
+                player.Kill();
+            }
         }
+    }
+
+
+    private void OnTriggerExit(Collider other)
+    {
+        if(knockoutOnExit){
+            Hittable hittable = other.gameObject.GetComponent<Hittable>();
+
+            if(hittable){
+                Player player = hittable.player;
+
+                player.Kill();
+            }
+        }
+        
     }
 }
